+++ backend/src/main/resources/db/migration/V4__create_menu_items_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE menu_items (
     id BIGSERIAL PRIMARY KEY,
     food_name VARCHAR(255),
     description VARCHAR(255),
-    price FLOAT,
+    price DECIMAL(10, 2),
     food_image VARCHAR(255),
     restaurant_id BIGINT,
     FOREIGN KEY (restaurant_id) REFERENCES restaurants (id)
